\name{XSBART-package}
\alias{XSBART-package}
\docType{package}
\title{
  XSBART: Accelerated soft Bayesian Additive Regression Trees
}
\description{
  A stochastic hill climbing algorithm for Soft Bayesian additive regression trees model. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Ran Hao 
Maintainer: Ran Hao <ranhao@163.shufe.edu.cn>
}
\references{
  XBART: Accelerated Bayesian Additive Regression Trees, AISTATS 2019.
  Soft BART: Bayesian regression tree ensembles that adapt to smoothness and sparsity,JRSSB 2018.
}
\keyword{ package }


